ported to ts, removed all js files, able to get data from external api
