+++ package.json
@@ -22,6 +22,10 @@
     "node-fetch": "^2.6.1"
   },
   "devDependencies": {
+    "@types/dotenv": "^8.2.0",
+    "@types/express": "^4.17.9",
+    "@types/node": "^14.14.10",
+    "@types/node-fetch": "^2.5.7",
     "typescript": "^4.1.2"
   }
 }
